+++ Formula/robotsmith.rb
@@ -16,7 +16,7 @@ class Robotsmith < Formula
   depends_on "go" => :build
 
   def install
-    system "go", "build", *std_go_args(ldflags: "-s -w -X main.version=v#{version}")
+    system "go", "build", *std_go_args(ldflags: "-X main.version=v#{version}")
   end
 
   test do